Move link import conflict check into schema-wide checks
